Extracted hudson-behavior.js find functions into 'find' module
